+++ src/Features/People/Core/Success/index.js
@@ -1,19 +1,16 @@
 import { useSelector } from "react-redux";
 import Pagination from "../../../../common/Pagination";
-import { fetchMoviesPending, selectPages } from "../../popularMoviesSlice";
-import Tile from "./Tile";
+import { fetchPeoplePending, selectPages } from "../../popularPeopleSlice";
+import Tile from "../../Tile";
 
 const Success = () => {
   const page = useSelector(selectPages);
-
   return (
     <>
       <Tile />
-      <Pagination page={page} fetchApi={fetchMoviesPending} />
+      <Pagination page={page} fetchApi={fetchPeoplePending} />
     </>
   )
-}
-  ;
+};
 
 export default Success;
-
