+++ src/Features/People/Tile/index.js
@@ -1,23 +1,14 @@
 import { PosterPeople, WrapperPeople, TileWrapperPeople, NamePeople } from "./styled"
 import { useSelector } from "react-redux";
-import { useEffect } from "react";
-import { useDispatch } from "react-redux";
-import { selectPeople, fetchPeoplePending } from "../popularPeopleSlice";
+import { selectPeople } from "../popularPeopleSlice";
 
+const Tile = () => {
 
-const PagePeople = () => {
-
-  const dispatch = useDispatch();
   const people = useSelector(selectPeople);
-
-  useEffect(() => {
-    dispatch((fetchPeoplePending(1)));
-  }, [dispatch])
 
   const posterPath = "https://image.tmdb.org/t/p/w300";
 
   return (
-
     <WrapperPeople>
       {people.map((person, id) => (
         <TileWrapperPeople key={id} >
@@ -28,4 +19,4 @@ const PagePeople = () => {
     </WrapperPeople>
   );
 }
-export default PagePeople;
+export default Tile;
