+++ src/Features/People/Tile/styled.js
