+++ src/Features/People/index.js
@@ -1,21 +1,25 @@
-import { useSelector } from "react-redux";
-import { selectPages } from "./popularPeopleSlice";
-import { fetchPeoplePending } from "./popularPeopleSlice";
+import { useEffect } from "react";
+import { useDispatch } from "react-redux";
 import { Container } from "../../common/Container/styled";
 import { Header } from "../../common/Header/styled";
-import Pagination from "../../common/Pagination";
-import PagePeople from "./PagePeople";
+import Core from "./Core";
+import { fetchPeoplePending } from "./popularPeopleSlice";
 
 const People = () => {
-  const page = useSelector(selectPages);
+
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch((fetchPeoplePending(1)));
+    // eslint-disable-next-line
+  }, [])
 
   return (
     <Container>
       <Header>
         Popular people
       </Header>
-      <PagePeople />
-      <Pagination page={page} fetchApi={fetchPeoplePending}/>
+      <Core />
     </Container>
   )
 };
