+++ src/common/Loading/styled.js
