+++ src/rootSaga.js
@@ -1,5 +1,5 @@
 import { all } from "redux-saga/effects";
-import { genresSaga } from "./Features/Movies/Tile/Genres/genresSaga";
+import { genresSaga } from "./Features/Movies/Core/Success/Tile/Genres/genresSaga";
 import { popularMoviesSaga } from "./Features/Movies/popularMoviesSaga";
 import { popularPeopleSaga } from "./Features/People/popularPeopleSaga"
 export default function* rootSaga() {
